jetson/requirements.txt
numpy
evdev
pyserial
